record.attachment overwritten by None value
I faced with an issue when attachment I send to rp_logger.info was empty in some cases. Going through the code, I noticed that wrap_makeRecord actually handles situations when "attachment" is in the "extra" parameter and then performs record.attachment = attachment (at the end of wrap_makeRecord -> makeRecord method). But when wrapper finishes, right in the next line code overwrites correct value of record.attachment with None value.